Fix testgui for Windows
QTemporaryFile::close() doesn't actually close the file behind the
scene. Because of this behavior running testgui on Windows results in
multiple failures:
1) testAutoreloadDatabase fails because Windows doesn't emit the file
changed event needed for auto-reload to work
2) testSave fails because of Access Denied error from QTemporaryFile
keeping it open

Fixing these two issues by creating a TemporaryFile class to act as a
proxy. On non-Windows platforms, all the functions go straight to the
QTemporaryFile counterparts. On Windows, the functions will use QFile
underneath the hood to make sure that once TemporaryFile::close() is
called, the file is really closed.