Move open db check from server to keyaccept dialog
To improve multi-db handling, we can move the checking for open
(unlocked) databases to the key accept dialog.

This means that even if the selected db is not open, the key can be
stored in another db as selected by the user.

This commit also adds a check for selected db's to the key accept dialog
so the dialog cannot be submitted if no db's to store the key are
selected.