Fix use of cmake_minimum_required
cmake_minimum_required() must be used before project(), otherwise it has no effect on internal cmake code that runs when project() is invoked. In particular CMP0025 does not default to NEW and CMake does not detect Clang features on Darwin, which is necessary to compile the project with the upstream (non Apple) compiler.