favicons: improve favicon fetching process
Numerous improvements:

* Eliminate dependency on libcurl. We already depend on Qt5Network, so
  why not use the HTTP fetching from there?

* Show a progress dialog when downloading the favicon. The main utility
  of this is giving the user the option to cancel a download attempt
  (e.g. if it's taking too long). Canceling will try the next fallback
  URL in the list.

* Try three different ways to obtain the favicon, in this order:

  * Direct to fully-qualified domain name
    (e.g. https://foo.bar.example.com/favicon.ico)
  * Direct to 2nd-level domain name
    (e.g. https://example.com/favicon.ico)
  * Google lookup for 2nd-level domain name (if enabled in settings)

I changed the Google lookup, because a match is more likely to be found
for the 2nd level domain than for the fully-qualified name.

Google's error behavior is strange. If it doesn't find a match, it
doesn't return an error. Instead, it returns a generic default icon,
which is not really the desired result. This also means that unless we
have some way to detect that we've received the generic icon, we can't
fall back to any alternatives.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>