Write CLI password prompt to stderr
Writing the prompt to stderr greatly helps with the use of keepassxc-cli
in automated scripts or as a password provider for programs like mutt or
borg.

This is done in accordance with POSIX that specifies that diagnostic
messages be sent to stderr.

This commit should be a first step in solving the following
issues: #831, #1221