Write CLI diagnostic messages to stderr.
As mentioned in 1dce5b09, POSIX specifies that all diagnostic and error
messages are to be sent to stderr, only meaningful data shall be output
to stdout (for example: attributes retrieved via keepassxc-cli show).

Stream variables were renamed using the <stream>TextStream convention
for consistency's sake.

It should be noted that stderr is supposed to be unbuffered, some calls
to flush() might not be useful anymore.