Add create command to keepassxc-cli
Command can be used to create a database using the command line. The
database can be secured by a key file and/or password. A key file can be
specified that will be created when it does not exist. The password will
be read from stdin, the same way as it's currently being done for
unlocking the database. When no key is specifed the function will fail.

Challenge response to unlock the database is not supported.

This functionality has been inspired by issue #2220, and solves point 1
of that issue.