Remove UrlFetchProgressDialog class
Per conversation in #2439, remove the progress dialog that appears when
downloading an entry's URL's favicon since (when working correctly) it
disappears before it can be read. When downloading icons from the button
located next to the URL text box, display a message panel that confirms
the download was a success.