Entry: delete m_tmpHistoryItem in destructor
Most of the time, `m_tmpHistoryItem` should be null by the time an
`Entry` is destroyed. However, if a caller ever calls `beginUpdate()`
without later calling `endUpdate()` -- perhaps because an exception was
throw in the meantime -- it may not be null.

This change avoids a memory leak in that case.

Found via https://lgtm.com/projects/g/keepassxreboot/keepassxc/alerts