Mention Nextcloud in addition to Dropbox
In the application settings dialog, the "Safely save database files"
checkbox warns that it may be incompatible with "Dropbox etc".
Since KeePassXC is all about privacy and open source, and as a nod
to a fellow open source project, this commit adds Nextcloud to the
warning message. The reason is that privacy-minded people, and thus
the main target group of KeePassXC, tend to switch away from
propriatary services like Dropbox and use open source alternatives,
of which Nextcloud seems to be among the most popular. If a
commercial system like Dropbox has reason to me mentioned in this
place, the same reason applies to Nextcloud as well. (Also, I've
used KeePassXC with Nextcloud long enough to confirm that the
option in question actually supports it perfectly well.)