Add extension ID for unpacked browser extension
This allows the browser extension to function correctly under Chromium
when it is loaded as an unpacked extension. This ID was generated for a
directory name of keepassxc by the Linux version of it. It may work for
other ports of Chromium but I have no way to test this right now.