FdoSecrets: Add confirmation dialog on secret access
Fixes #3837

The 'comm' of a process can be changed without any special permissions by the process itself. Use the 'exe' link for a more reliable identification of a service user. This link points to the actual executable. The real identity might still be hidden by using user/mount namespaces or by prctl(PR_SET_DUMPABLE, 0).

fdosecrets: track clients in dedicated Connection class

As per specification it is allowed that a client creates multiple Sessions objects. Previously only the last created Session per client was remembered. Now a client can create any number of Sessions. This is
done by creating a single Connection class instance on the first request of a client. The Session objects are childs of the Connection and will be garbage collected if the client disconnects from the bus.

At the same time the ability to forcefully "disconnect" a client is taken over by the Connection class too. Whenever a client is thrown out it is no longer able to use newly created sessions.

fdosecrets: add access confirmation setting

Add a knob to request confirmation when a password is access through the FDO secrets service. This confirmation is enabled by default but the
user can disable it to regain the old behaviour.

fdosecrets: implement confirmation on secret access

If confirmations are enabled in the database setting the items are locked by default for a client. On the first access the user is able to unlock this item or all items for a client. If confirmations are disabled the items are unlocked as soon as the database is unlocked.

The choice of the user is remembered for the client connection. The DBus specification guarantees that the client address is not reused during the lifetime of a DBus session. So as soon as the client process
terminates a new confirmation will be required the next time. As long as the client service is alive it can be authorized for the whole session for all items by choosing "Yes to all". By choosing "Yes" the access is
granted only for the currently requested item.

The implementation does not block the DBus handler but relies on the prompt objects of the API. Unfortunately there is no way in QtDBus to
know who reads a property. You can only get the caller identity on slot calls. This has the unfortunate consequence that we cannot honestly report the Locked property of the items. So even when an application has been given access to a certain item it will still read the Locked property as true.