Replace entry password in database with clipboard
This adds a new context menu item and hotkey to replace the password
with the contents of the clipboard.

For my use this adds a convenience feature where I often have to copy a
new password from an alternative system and update the database with the
new copy for different entries throughout the day.

Signed-off-by: ed neville <ed@s5h.net>