icon purge: restore icon selection if possible
when purging unused custom icons, restore the selection to be the custom
icon that was selected in the moment when clicking the purge button. no
action required if the previous selected icon was from the default icon
set.

as a fallback, the custom icon associated with the entry or group is
selected after purging in case the last selected icon was purged. that
happens when changing the selecting to a custom icon that would then be
in use only by the current entry or group if the entry/group was saved.
since the icon is not yet recorded as the entry or group icon, it is
purged.