remove icon deletion/purge from edit icon widget
this functionality is now available in the new database settings icons
widget. managing custom icons as part of the database should not
happend when editing an entry or group, but should be a separate
feature:

- no need to elaborately select a suitable icon after deleting a single
icon (that has to be selected first, but is not selected as the new
icon for the currelty edited entry or group).

- purging unused icons becomes complicated when the user selects a custom
icon to be the new icon for an entry/group, but that icon is currently
not in use. that icon is not yet associated with the entry/group and
hence is removed when purging unused icons. assuming the newly selected
icon is the new icon for the entry/group is not an option: the user can
select another icon after purging. a similar problem arises when the
active icon for an entry/group is only used by that entry/group and
would become unused when selecting another icon.